Refactor: Settings models moved to modules subpackage
