+++ bpod_rig/config/models/__init__.py
@@ -1,0 +1,1 @@
+"""Pydantic models for Bpod-rig settings"""
